requirements.txt
akshare>=1.14.0
anthropic>=0.40.0
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
